Finishing serialization/deserialization

+++ Control/GameState.cs
@@ -7,6 +7,7 @@ using Microsoft.Xna.Framework.Audio;
 using Microsoft.Xna.Framework.Content;
 using Microsoft.Xna.Framework.Graphics;
 using Microsoft.Xna.Framework.Input;
+using Microsoft.Xna.Framework.Media;
 using MonoGame.Extended;
 using MonoGame.Extended.Tiled;
 using Newtonsoft.Json;
@@ -18,9 +19,9 @@ using static GameAttempt1.Utilities.GameUtilities;
 
 namespace GameAttempt1.Control
 {
-    [Serializable]
     public sealed class GameState : State
     {
+        #region Fields
         private readonly Player _player;
         private readonly List<Component> _pauseComponents;
         private Texture2D _playerTextures;
@@ -28,6 +29,8 @@ namespace GameAttempt1.Control
         private SoundEffect _jumpSoundEffect;
         private Button _pauseButton;
         private Camera _camera;
+        #endregion
+        #region Constructors
         public GameState(ContentManager contentManager, TwoDPlatformer game, GraphicsDevice graphicsDevice, LevelController levelController)
             : base(contentManager, game, graphicsDevice)
         {
@@ -41,9 +44,19 @@ namespace GameAttempt1.Control
             _levelController.Current.PlaySong();
             LoadComponents();
         }
-
-
-
+        public GameState(ContentManager contentManager, TwoDPlatformer game, GraphicsDevice graphicsDevice, LevelController levelController, PlayerData playerData)
+            : base(contentManager, game, graphicsDevice)
+        {
+            _pauseComponents = new List<Component>();
+            _playerTextures = _contentManager.Load<Texture2D>("Sprites/Tuxedo");
+            _player = new Player(playerData);
+            _player.LoadTexture(_playerTextures);
+            _levelController = levelController;
+            _levelController.Current.PlaySong();
+            LoadComponents();
+        }
+        #endregion
+        #region Content-Loading Methods
         public void LoadComponents()
         {
             _camera = new Camera();
@@ -54,6 +67,10 @@ namespace GameAttempt1.Control
 
             var buttonFont = _contentManager.Load<SpriteFont>("ButtonFont");
             var buttonTexture = _contentManager.Load<Texture2D>("Button");
+            var sliderTexture = _contentManager.Load<Texture2D>("Slider");
+            var sliderScrollTexture = _contentManager.Load<Texture2D>("SliderScroll");
+            var volumeFont = _contentManager.Load<SpriteFont>("VolSliderDescriptionFont");
+
             var saveButton = new Button(buttonTexture, buttonFont,
                 new Vector2(MENU_X_COORDINATE, MENU_Y_COORDINATE),
                 "Save");
@@ -62,13 +79,26 @@ namespace GameAttempt1.Control
                 new Vector2(MENU_X_COORDINATE, MENU_Y_COORDINATE + MENU_OFFSET),
                 "Exit");
             mainMenuButton.ButtonPress += ExitGame_OnClick;
+
+            var volumeCounter = new Counter<float>(volumeFont, new Vector2(MENU_X_COORDINATE,
+            MENU_Y_COORDINATE + 3 * MENU_OFFSET), "Volume", 100, () => (float)Math.Round(MediaPlayer.Volume * 100));
+            var volumeSlider = new Slider(sliderTexture, sliderScrollTexture, new Vector2(MENU_X_COORDINATE,
+                MENU_Y_COORDINATE + 4 * MENU_OFFSET));
+            volumeSlider.SliderScroll += VolumeSlider_VolumeChanged;
+
             _jumpSoundEffect = _contentManager.Load<SoundEffect>("Sounds/jumppp11");
             _player.Radio += PlaySoundEffect_OnJump;
-            _pauseComponents.Add(saveButton, mainMenuButton);
+            _pauseComponents.Add(saveButton, mainMenuButton, volumeCounter, volumeSlider);
         }
+        #endregion
+        #region GameState Events
         public void PlaySoundEffect_OnJump(object sender, EventArgs e)
         {
             _jumpSoundEffect.Play();
+        }
+        private static void VolumeSlider_VolumeChanged(object sender, float volume)
+        {
+            MediaPlayer.Volume = volume;
         }
         public void ExitGame_OnClick(object sender, EventArgs e)
         {
@@ -82,24 +112,37 @@ namespace GameAttempt1.Control
             data.State = _player.State;
             data.Direction = _player.Direction;
             data.Velocity = _player.Velocity;
+            var gameData = new GameData();
+            gameData.PlayerData = data;
+            gameData.Level = _levelController.Index;
 
-            string jsonData = JsonConvert.SerializeObject(data, Formatting.Indented);
+            string jsonData = JsonConvert.SerializeObject(gameData, Formatting.Indented);
             int fileNum = 1;
-            var filename = $"../../../Saves/save{fileNum}";
-            while (File.Exists(filename))
+            var filename = $"{DIR_PATH_RELATIVE}Saves/save{fileNum}";
+            while (File.Exists(filename) && fileNum < 4)
             {
                 fileNum++;
-                filename = $"../../../Saves/save{fileNum}";
+                filename = $"{DIR_PATH_RELATIVE}Saves/save{fileNum}";
             }
+            if(!File.Exists(filename))
+            {
+                File.WriteAllText(filename, jsonData);
+                return;
+            }
+            File.Delete($"{DIR_PATH_RELATIVE}Saves/save1");
+            File.Move($"{DIR_PATH_RELATIVE}Saves/save2", $"{DIR_PATH_RELATIVE}Saves/save1");
+            File.Move($"{DIR_PATH_RELATIVE}Saves/save3", $"{DIR_PATH_RELATIVE}Saves/save2");
+            File.Move($"{DIR_PATH_RELATIVE}Saves/save4", $"{DIR_PATH_RELATIVE}Saves/save3");
             File.WriteAllText(filename, jsonData);
+
         }
         public void PauseGame_OnPress(object sender, EventArgs e)
         {
             _levelController.Current.Pause();
             _player.Pause();
         }
-
-
+        #endregion
+        #region State Methods
         public override void Draw(GameTime gameTime, SpriteBatch spriteBatch)
         {
             spriteBatch.Begin(transformMatrix: _camera.Transform);
@@ -119,6 +162,38 @@ namespace GameAttempt1.Control
             }
             spriteBatch.End();
         }
+        public override void Update(GameTime gameTime)
+        {
+            PlayerExternalUpdate();
+            _player.Update(gameTime);
+            if (_player.Direction == GameDirection.Right
+                && _player.Position.X > _levelController.Current.GlobalBounds.X - PLAYER_WIDTH * 2) _player.Velocity.X = 0;
+            _camera.Follow(_player, _graphicsDevice);
+            if (IsPlayerDead())
+            {
+                _levelController.Current.ResetSong();
+                _player.Reset();
+            }
+            _pauseButton.Update(gameTime);
+            _levelController.Update(gameTime);
+            foreach (var entity in _levelController.Current.Entities)
+            {
+                if (_levelController.PlayerEntityCollision(_player, entity))
+                {
+                    if (entity.GetType() == typeof(Coin))
+                    {
+                        ((Coin)entity).Picked = true;
+                    }
+                }
+            }
+            if (_player.State == PlayerState.Paused)
+            {
+                _pauseComponents.ForEach(c => c.Update(gameTime));
+            }
+
+        }
+        #endregion
+        #region Debug Methods
         public void DrawDebug(GraphicsDevice gd, SpriteBatch spriteBatch)
         {
             var textures = _levelController.Current.TiledMap.GetLayer<TiledMapObjectLayer>("Platforms").Objects;
@@ -145,6 +220,8 @@ namespace GameAttempt1.Control
                 spriteBatch.Draw(rectangleTexture2D, t.Position, Color.Red);
             }
         }
+        #endregion
+        #region External Player Related Methods
         public void PlayerExternalUpdate()
         {
             var platforms = _levelController.Current.TiledMap.GetLayer<TiledMapObjectLayer>("Platforms")
@@ -179,40 +256,10 @@ namespace GameAttempt1.Control
             }
         }
 
-
-        public override void Update(GameTime gameTime)
-        {
-            PlayerExternalUpdate();
-            _player.Update(gameTime);
-            if (_player.Direction == GameDirection.Right && _player.Position.X > _levelController.Current.GlobalBounds.X - PLAYER_WIDTH * 2) _player.Velocity.X = 0;
-            _camera.Follow(_player, _graphicsDevice);
-            if (IsPlayerDead())
-            {
-                _levelController.Current.ResetSong();
-                _player.Reset();
-            }
-            _pauseButton.Update(gameTime);
-            _levelController.Update(gameTime);
-            foreach (var entity in _levelController.Current.Entities)
-            {
-                if(_levelController.PlayerEntityCollision(_player, entity))
-                {
-                    if(entity.GetType() == typeof(Coin))
-                    {
-                        ((Coin)entity).Picked = true;
-                    }
-                }
-            }
-            if (_player.State == PlayerState.Paused)
-            {
-                _pauseComponents.ForEach(c => c.Update(gameTime));
-            }
-
-        }
-
         public bool IsPlayerDead()
         {
             return (_player.Position.Y > _graphicsDevice.Viewport.Height);
         }
+        #endregion
     }
 }
